chore(web): add render optimization tooling
Keep React Compiler adoption safe and measurable by defaulting to annotation mode and wiring in a dev-only React Scan workflow for validation.
