+++ package.json
@@ -4,6 +4,10 @@
   "private": true,
   "scripts": {
     "dev": "next dev --turbopack",
+    "dev:scan": "NEXT_PUBLIC_DELTA_REACT_SCAN=true next dev --turbopack",
+    "dev:scan:unnecessary": "NEXT_PUBLIC_DELTA_REACT_SCAN=true NEXT_PUBLIC_DELTA_REACT_SCAN_TRACK_UNNECESSARY=true next dev --turbopack",
+    "dev:compiler": "DELTA_REACT_COMPILER_MODE=infer next dev --turbopack",
+    "dev:compiler:scan": "DELTA_REACT_COMPILER_MODE=infer NEXT_PUBLIC_DELTA_REACT_SCAN=true next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "lint": "biome check .",
@@ -60,8 +64,10 @@
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "@vitest/coverage-v8": "^4.1.0",
+    "babel-plugin-react-compiler": "^1.0.0",
     "drizzle-kit": "^0.31.10",
     "react-grab": "^0.1.29",
+    "react-scan": "^0.5.3",
     "tailwindcss": "^4",
     "tsx": "^4.21.0",
     "typescript": "^5",
