+++ src/lib/react-client.ts
@@ -1,0 +1,1 @@
+export { useEffect } from "react";
